+++ Source/Public/Compare-ObjectGraph.ps1
@@ -56,7 +56,6 @@
 
 .PARAMETER MaxDepth
     The maximal depth to recursively compare each embedded property (default: 10).
-
 #>
 function Compare-ObjectGraph {
     [CmdletBinding()] param(